switched convert and format commands to be consistent with strs [skip ci]
