+++ lusSTR/data/snp_config.yaml
@@ -15,3 +15,4 @@ nofilter: False ## True/False if no filtering is desired; if False, will remove 
 strand: "forward" ## forward/uas; indicates which oritentation to report the alleles for the ForenSeq SNPs; uas indicates the orientation as reported by the UAS or the forward strand
 references: "" ## list IDs of the samples to be run as references in EFM
 separate: false ## True/False; if want to separate samples into individual files for use in EFM
+thresh: 0.03 ## Analytical threshold value
